whisperforge/requirements.txt
# WhisperForge Dependencies

# Core Web Framework
streamlit>=1.28.0

# AI/ML Libraries
openai>=1.3.0
anthropic>=0.7.0

# Audio Processing
pydub>=0.25.0

# Database & Backend
supabase>=2.0.0
python-dotenv>=1.0.0

# Authentication & OAuth
streamlit-oauth>=0.1.0
httpx>=0.24.0
urllib3>=2.0.0

# Utilities
requests>=2.31.0
python-jwt>=4.0.0
cryptography>=41.0.0

# File Processing
python-multipart>=0.0.6